requirements.txt
Flask==3.0.0
pycryptodome==3.20.0
Werkzeug==3.0.1
